tidy up code
